LayoutTests/js/kde/var_decl_init-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS varInFunction() is true
PASS varInVarList() is true
PASS varListOrder() is true
PASS varInBlock() is true
PASS varInIf() is true
PASS varInElse() is true
PASS varInDoWhile() is true
PASS varInWhile() is true
PASS varInFor() is true
PASS varInForIn() is true
PASS varInWith() is true
PASS varInCase() is true
PASS varInLabel() is true
PASS varInCatch() is true
PASS varInFinally() is true
PASS varInTry() is true
PASS varInForInitExpr() is true
PASS varInSubFunction() is true
PASS varGlobal is 1
PASS overrideVar is 1
PASS overrideVar2 is 2
PASS successfullyParsed is true

TEST COMPLETE
